# Conflicts:
#	app/src/main/res/layout/er_fragment_employeemanage.xml

+++ app/src/main/java/com/example/albatong/data/UserSalary.kt
@@ -1,4 +1,4 @@
 package com.example.albatong.data
 
-data class UserSalary(var storeName: String, var totalMinutes: Int, var monthlySalary: Float) {
+data class UserSalary(var storeName: String, var totalMinutes: Int, var monthlySalary: Int) {
 }
